Add option to allow repeated references of jobs to be shown on the dashboard
Enabling the option will allow the user to refer the same job multiple
times in the Explicit Order text box and have those multiple references
to be respected in the actual dashboard.
This is change was motivated by organisational purposes of the
dashboard.

Unit/acceptance tests not created yet, but retro-compatibility with
existing ones was kept.